refactor: relocate fix_local_permissions out of deploy — fully severs extension build -> deploy
fix_local_permissions normalizes group read/write on a local component checkout
before it's archived (some tools create restrictive 600 perms). It's a local-fs
utility (deps: engine::shell::quote_path, defaults, log_status, Command) with no
SSH/deploy machinery — only extension/build called it, yet it lived under
deploy::permissions, forcing the build path to import crate::deploy.

Relocate it to a core-level crate::local_permissions module. The SSH-based
fix_deployed_permissions stays in deploy::permissions (deploy-only). This removes
the LAST crate::deploy reference from the extension subsystem — extension no
longer depends on deploy at all (0 files, was 1).

Combined with the tag-gap relocation (#8792), the extension build path's two
deploy edges are both severed.

Also completes 2 pre-existing broken cli test constructors: AgentTaskPromotion
Options gained a required candidate_ref field on main but agent_task_promotion/
tests.rs still omitted it (cargo check --workspace --tests was already red on
clean main). Added candidate_ref: None. Unrelated to this change.

Refs #8425
