+++ crates/homeboy-core/src/agent_task_promotion/tests.rs
@@ -1248,6 +1248,7 @@ fn empty_patch_runs_public_and_private_gates_against_pinned_candidate() {
             source_worktree_path: Some(repo.clone()),
             base_ref: None,
             task_base_sha: None,
+            candidate_ref: None,
             to_worktree: "repo@candidate".to_string(),
             task_id: None,
             artifact_id: None,
@@ -1303,6 +1304,7 @@ fn empty_patch_failing_gate_is_reported_against_pinned_candidate() {
             source_worktree_path: Some(repo.clone()),
             base_ref: None,
             task_base_sha: None,
+            candidate_ref: None,
             to_worktree: "repo@candidate".to_string(),
             task_id: None,
             artifact_id: None,
